Move AddToPlaylist context menu entry class
