+++ OsuPlayer.Data/OsuPlayer/Classes/AddToPlaylistContextMenuEntry.cs
@@ -1,0 +1,13 @@
+﻿namespace OsuPlayer.Data.OsuPlayer.Classes;
+
+public class AddToPlaylistContextMenuEntry
+{
+    public string Name { get; set; }
+    public Action<string> Action { get; set; }
+
+    public AddToPlaylistContextMenuEntry(string name, Action<string> action)
+    {
+        Name = name;
+        Action = action;
+    }
+}
